THIS WORKS.
It could be more optimized but it now creates all the systems folders in the $BACKUP_DIR
Current Bugs:

It still will create a $BACKUP_DIR folder inside of itself but I am working on a fix
Dreamcast saves are '.bin' files. Will build in special logic as to not backup every bin file since that's also a filetype for other system ROMs.